Move ExampleWiki to ExampleTopic, add .gitignore.
There was some potential confusion as to the purpose of
$MWK/ExampleWiki.

I changed it to $MWK/ExampleTopic,
- to promote the idea of topic directories at the top of the wiki,
- to reduce the likelihood that a new user would put their own files in
  ExampleTopic,
- thereby making it easier for new users to eventually remove
  ExampleTopic.

I added a .gitignore file.
I was originally using .git/info/exclude, not realizing that that file
is just for me, *and* that I was preventing the zero-length media files
in ExampleTopic/Books/* from being committed.

.gitignore ignores .html, .pdf and .txt,
and specifically unignores the zero-length media files under
ExampleTopic/Books.