Update README.rst for ExampleTopic reorg.
In the previous commit, ExampleWiki was changed to ExampleTopic,
with links updated accordingly.
This update to README.rst reflects that change,
and emphasizes the nature of ExampleTopic as a topic,
not a complete demo wiki.

uppity.png was re-shot, to show the new ExampleTopic name.

Minor change to makefile, to silence "mwk clean".