Fix Issue #7: make hangs when no meta.json exists.
makefile:
makefile is written so that if a target's sources don't exist,
the target's make variable is not set, and so that target is not
attempted.

CATA was brute-force set to catalog.json, regardless.
If no meta.json existed,
make hung making that target, as no sources were found.

Set CATA to empty unless meta.json is found, fixed.